renamed component [Recipe -> Recipes]
